feat: verifyToken을 유틸 함수로 이동하고 필요한 페이지에서만 처리
