+++ src/frontend/pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from 'next/app';
-import { useRouter } from 'next/router';
+import { NextPage, NextPageContext } from 'next';
 import { ThemeProvider } from 'styled-components';
 
 import { GlobalStyle } from 'styles/global-styles';
@@ -8,37 +8,15 @@ import { theme } from 'styles/theme';
 import wrapper from 'modules/configureStore';
 import Layout from 'templates/Layout';
 
-import { parseToken } from 'util/parseToken';
-import { reIssueTokenAPI } from 'api/auth/api';
-
 import 'styles/globals.css';
 import 'styles/font.css';
 
+// interface IAppProps {
+//   Component: NextPage;
+//   ctx: NextPageContext;
+// }
+
 function MyApp({ Component, pageProps }: AppProps) {
-  const verifyToken = async () => {
-    const accessToken = localStorage.getItem('accessToken');
-    const refreshToken = localStorage.getItem('refreshToken');
-
-    //토큰 만료시간 1분전
-    const accessTokenTime = parseToken(accessToken!).exp * 1000 - 60000;
-    const refreshTokenTime = parseToken(refreshToken!).exp * 1000 - 60000;
-
-    const router = useRouter();
-
-    //refreshToken이 만료되었다면 로그인 페이지로 이동
-    if (refreshTokenTime < Date.now()) {
-      router.push('/signin');
-    }
-    //accessToken이 만료되었다면 accessToken 재발급하여 다시 저장
-    else if (accessTokenTime < Date.now()) {
-      if (accessToken && refreshToken) {
-        const newAccessToken = (await reIssueTokenAPI({ accessToken, refreshToken, isUpdated: false })).data
-          .accessToken;
-        localStorage.setItem('accessToken', newAccessToken);
-      }
-    }
-  };
-
   return (
     //themeprovider를 이용해서 미리 정의한 테마 타입들을 뷰 단에서 사용
     <ThemeProvider theme={theme}>
@@ -49,6 +27,20 @@ function MyApp({ Component, pageProps }: AppProps) {
     </ThemeProvider>
   );
 }
+
+// 여기서 실행 안됨 일단은
+// MyApp.getInitialProps = async ({ Component, ctx }: IAppProps) => {
+//   let pageProps = {};
+
+//   // 실행하고자 하는 component에 getInitialprops가 있으면 실행하여 props를 받아올 수 있다.
+//   if (Component.getInitialProps) {
+//     pageProps = await Component.getInitialProps(ctx);
+//   }
+
+//   return {
+//     pageProps,
+//   };
+// };
 
 // export default wrapper.withRedux(withReduxSaga(MyApp));
 export default wrapper.withRedux(MyApp);
